chore: split FederationPeer struct into separate file
