+++ fedimint-api-client/src/api/mod.rs
@@ -3,7 +3,7 @@ use std::fmt::{self, Debug, Display};
 use std::num::NonZeroUsize;
 use std::pin::Pin;
 use std::sync::Arc;
-use std::time::{Duration, SystemTime};
+use std::time::Duration;
 use std::{cmp, result};
 
 use anyhow::anyhow;
@@ -36,12 +36,9 @@ use fedimint_core::module::audit::AuditSummary;
 use fedimint_core::module::registry::ModuleDecoderRegistry;
 use fedimint_core::module::{ApiAuth, ApiRequestErased, ApiVersion, SerdeModuleEncoding};
 use fedimint_core::session_outcome::{AcceptedItem, SessionOutcome, SessionStatus};
-use fedimint_core::task::jit::JitTryAnyhow;
 use fedimint_core::task::{MaybeSend, MaybeSync};
-use fedimint_core::time::now;
 use fedimint_core::transaction::{SerdeTransaction, Transaction, TransactionSubmissionOutcome};
-use fedimint_core::util::backon::BackoffBuilder;
-use fedimint_core::util::{backon, SafeUrl};
+use fedimint_core::util::SafeUrl;
 use fedimint_core::{
     apply, async_trait_maybe_send, dyn_newtype_define, runtime, NumPeersExt, OutPoint, PeerId,
     TransactionId,
@@ -61,10 +58,14 @@ use jsonrpsee_ws_client::{WsClient, WsClientBuilder};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use thiserror::Error;
-use tokio::sync::{Mutex, OnceCell, RwLock};
+use tokio::sync::OnceCell;
 use tracing::{debug, error, instrument, trace, warn};
 
 use crate::query::{FilterMapThreshold, QueryStep, QueryStrategy, ThresholdConsensus};
+
+mod federation_peer_client;
+
+use federation_peer_client::FederationPeer;
 
 pub type PeerResult<T> = Result<T, PeerError>;
 pub type JsonRpcResult<T> = Result<T, JsonRpcClientError>;
@@ -1118,151 +1119,6 @@ pub struct WsFederationApi<C = WsClient> {
     module_id: Option<ModuleInstanceId>,
 }
 
-/// Connection state shared/preserved between [`FederationPeerClient`] and the
-/// Jit tasks it spawns.
-#[derive(Debug)]
-struct FederationPeerClientConnectionState {
-    /// Last time a connection attempt was made, or `None` if no attempt has
-    /// been made yet.
-    last_connection_attempt_or: Option<SystemTime>,
-    connection_backoff: backon::FibonacciBackoff,
-}
-
-impl FederationPeerClientConnectionState {
-    const MIN_BACKOFF: Duration = Duration::from_millis(100);
-    const MAX_BACKOFF: Duration = Duration::from_secs(5);
-
-    pub fn new() -> Self {
-        Self {
-            last_connection_attempt_or: None,
-            connection_backoff: Self::new_backoff(),
-        }
-    }
-
-    /// Wait (if needed) before reconnection attempt based on number of previous
-    /// attempts and update reconnection stats.
-    async fn wait(&mut self) {
-        let desired_timeout = self.connection_backoff.next().unwrap_or(Self::MAX_BACKOFF);
-        let since_last_connect = match self.last_connection_attempt_or {
-            Some(last) => now().duration_since(last).unwrap_or_default(),
-            None => Duration::ZERO,
-        };
-
-        let sleep_duration = desired_timeout.saturating_sub(since_last_connect);
-        if Duration::ZERO < sleep_duration {
-            debug!(
-                target: LOG_CLIENT_NET_API,
-                duration_ms=sleep_duration.as_millis(),
-                "Waiting before reconnecting");
-        }
-        fedimint_core::runtime::sleep(sleep_duration).await;
-
-        self.last_connection_attempt_or = Some(now());
-    }
-
-    fn reset(&mut self) {
-        *self = Self::new();
-    }
-
-    fn new_backoff() -> backon::FibonacciBackoff {
-        backon::FibonacciBuilder::default()
-            .with_min_delay(Self::MIN_BACKOFF)
-            .with_max_delay(Self::MAX_BACKOFF)
-            .build()
-    }
-}
-
-/// The client in [`FederationPeer`], that takes care of reconnecting by
-/// starting background Jit task
-#[derive(Debug)]
-struct FederationPeerClient<C> {
-    client: JitTryAnyhow<C>,
-    connection_state: Arc<tokio::sync::Mutex<FederationPeerClientConnectionState>>,
-}
-
-impl<C> FederationPeerClient<C>
-where
-    C: JsonRpcClient + 'static,
-{
-    pub fn new(peer_id: PeerId, url: SafeUrl, api_secret: Option<String>) -> Self {
-        let connection_state = Arc::new(tokio::sync::Mutex::new(
-            FederationPeerClientConnectionState::new(),
-        ));
-
-        Self {
-            client: Self::new_jit_client(peer_id, url, api_secret, connection_state.clone()),
-            connection_state,
-        }
-    }
-
-    fn new_jit_client(
-        peer_id: PeerId,
-        url: SafeUrl,
-        api_secret: Option<String>,
-        connection_state: Arc<Mutex<FederationPeerClientConnectionState>>,
-    ) -> JitTryAnyhow<C> {
-        JitTryAnyhow::new_try(move || async move {
-            Self::wait(&peer_id, &url, &connection_state).await;
-
-            let res = C::connect(&url, api_secret).await;
-
-            match &res {
-                Ok(_) => {
-                    connection_state.lock().await.reset();
-                    debug!(
-                            target: LOG_CLIENT_NET_API,
-                            peer_id = %peer_id,
-                            url = %url,
-                            "Connected to peer");
-                }
-                Err(err) => {
-                    debug!(
-                            target: LOG_CLIENT_NET_API,
-                            peer_id = %peer_id,
-                            url = %url,
-                            %err, "Unable to connect to peer");
-                }
-            }
-            Ok(res?)
-        })
-    }
-
-    pub fn reconnect(&mut self, peer_id: PeerId, url: SafeUrl, api_secret: Option<String>) {
-        self.client = Self::new_jit_client(peer_id, url, api_secret, self.connection_state.clone());
-    }
-
-    async fn wait(
-        peer_id: &PeerId,
-        url: &SafeUrl,
-        connection_state: &Arc<Mutex<FederationPeerClientConnectionState>>,
-    ) {
-        let mut connection_state_guard = connection_state.lock().await;
-
-        if connection_state_guard.last_connection_attempt_or.is_none() {
-            debug!(
-                target: LOG_CLIENT_NET_API,
-                peer_id = %peer_id,
-                url = %url,
-                "Connecting to peer...");
-        } else {
-            debug!(
-                target: LOG_CLIENT_NET_API,
-                peer_id = %peer_id,
-                url = %url,
-                "Retrying connecting to peer...");
-        }
-
-        connection_state_guard.wait().await;
-    }
-}
-
-#[derive(Debug)]
-struct FederationPeer<C> {
-    url: SafeUrl,
-    peer_id: PeerId,
-    api_secret: Option<String>,
-    client: RwLock<FederationPeerClient<C>>,
-}
 impl<C: JsonRpcClient + Debug + 'static> IModuleFederationApi for WsFederationApi<C> {}
 
 /// Implementation of API calls over websockets
@@ -1422,16 +1278,7 @@ where
                         );
                         assert!(url.host().is_some(), "API client requires a target host");
 
-                        FederationPeer {
-                            peer_id,
-                            client: RwLock::new(FederationPeerClient::new(
-                                peer_id,
-                                url.clone(),
-                                api_secret.clone(),
-                            )),
-                            url,
-                            api_secret: api_secret.clone(),
-                        }
+                        FederationPeer::new(url, peer_id, api_secret.clone())
                     })
                     .collect(),
             ),
